forgot to rename the app to notebook_app in the go module file
