+++ cmd/app/auth/register.go
@@ -9,12 +9,12 @@ import (
 	
 	"errors"
 	
-	"go-web-app-experiment/cmd/app/notebook_db"
+	"notebook_app/cmd/app/notebook_db"
 
-	"go-web-app-experiment/cmd/app/form"
+	"notebook_app/cmd/app/form"
 )
 
-func register_new_user(db *gorm.DB, email string, username string, password string, confirm string) error {
+func RegisterNewUser(db *gorm.DB, email string, username string, password string, confirm string) error {
 	/* Check if email is already taken */
 	var user1 = &notebook_db.User{}
 
@@ -68,7 +68,7 @@ func Register(db *gorm.DB) gin.HandlerFunc {
 		var confirm string = form.GetFormValue(c, "confirm") 
 
 		// Register new user
-		err := register_new_user(db, email, username, password, confirm)
+		err := RegisterNewUser(db, email, username, password, confirm)
 
 		/* Check if user registration is successful */
 		if err == nil {
