+++ cmd/app/notes/view_notes_page.go
@@ -7,9 +7,9 @@ import (
 
 	"gorm.io/gorm"
 
-	"go-web-app-experiment/cmd/app/notebook_db"
+	"notebook_app/cmd/app/notebook_db"
 
-	"go-web-app-experiment/cmd/app/user_session"
+	"notebook_app/cmd/app/user_session"
 )
 
 /* Get Requests */
