+++ cmd/app/tests/auth_test.go
@@ -1,0 +1,28 @@
+package tests
+
+import (
+	//"net/http"
+	"testing"
+
+	//"github.com/stretchr/testify/assert"
+
+	//"notebook_app/cmd/app/auth"
+)
+
+func TestRegistration(t *testing.T) {
+	// Change directory
+	ChangetoRepoRootDirectory()
+
+	// Setup App
+	//var r, db = RunApp()
+
+	//auth.RegisterNewUser()
+	
+
+	// test home page
+	//request, _ := http.NewRequest("GET", "/register", nil)
+	//r.ServeHTTP(write, request)
+	//assert.Equal(t, 200, write.Code)
+}
+
+
